chore(appium): Migrate the entire schema from JavaScript to TypeScript
