+++ packages/appium/lib/schema/index.ts
